+++ lib/src/api/api_account_routes.dart
@@ -266,9 +266,10 @@ class AccountRoutesApi {
     if (response.statusCode >= 400) {
       throw ApiException(response.statusCode, response.body);
     } else if (response.body != null) {
-      final List resp = (_apiClient.deserialize(response.body, 'List<Transaction>'))
-          .map((item) => item)
-          .toList();
+      final List resp =
+          (_apiClient.deserialize(response.body, 'List<Transaction>'))
+              .map((item) => item)
+              .toList();
       return resp.map(_deserializeDTO).toList();
     } else {
       return null;
@@ -332,9 +333,10 @@ class AccountRoutesApi {
     if (response.statusCode >= 400) {
       throw ApiException(response.statusCode, response.body);
     } else if (response.body != null) {
-      final List resp = (_apiClient.deserialize(response.body, 'List<Transaction>'))
-          .map((item) => item)
-          .toList();
+      final List resp =
+          (_apiClient.deserialize(response.body, 'List<Transaction>'))
+              .map((item) => item)
+              .toList();
       return resp.map(_deserializeDTO).toList();
     } else {
       return null;
@@ -397,9 +399,10 @@ class AccountRoutesApi {
     if (response.statusCode >= 400) {
       throw ApiException(response.statusCode, response.body);
     } else if (response.body != null) {
-      final List resp = (_apiClient.deserialize(response.body, 'List<Transaction>'))
-          .map((item) => item)
-          .toList();
+      final List resp =
+          (_apiClient.deserialize(response.body, 'List<Transaction>'))
+              .map((item) => item)
+              .toList();
       return resp.map(_deserializeDTO).toList();
     } else {
       return null;
@@ -462,9 +465,10 @@ class AccountRoutesApi {
     if (response.statusCode >= 400) {
       throw ApiException(response.statusCode, response.body);
     } else if (response.body != null) {
-      final List resp = (_apiClient.deserialize(response.body, 'List<Transaction>'))
-          .map((item) => item)
-          .toList();
+      final List resp =
+          (_apiClient.deserialize(response.body, 'List<Transaction>'))
+              .map((item) => item)
+              .toList();
       return resp.map(_deserializeDTO).toList();
     } else {
       return null;
@@ -527,9 +531,10 @@ class AccountRoutesApi {
     if (response.statusCode >= 400) {
       throw ApiException(response.statusCode, response.body);
     } else if (response.body != null) {
-      final List resp = (_apiClient.deserialize(response.body, 'List<Transaction>'))
-          .map((item) => item)
-          .toList();
+      final List resp =
+          (_apiClient.deserialize(response.body, 'List<Transaction>'))
+              .map((item) => item)
+              .toList();
       return resp.map(_deserializeDTO).toList();
     } else {
       return null;